Fix : UpdateChannelModal의 채널 삭제 버튼 위치 수정
