+++ frontend/src/components/Main/ContentsSection/MessageSection.tsx
@@ -3,12 +3,12 @@ import styled from 'styled-components';
 
 import { MainStoreContext } from '../MainStore';
 import { MessageData, MessageListInfo, MessageRequestBody } from '../../../types/message';
-import noInfoImg from '../../../assets/hmm.gif';
 
 import { TextareaDiv, MessageTextarea, messageInnerElement, onKeyDownMessageTextarea } from './ContentsSectionCommon';
 import { User } from '../../../types/user';
 import ChannelEntity from '../../../types/channel';
 import UserListModal from './UserListModal';
+import NotFoundChannel from './NotFoundChannel';
 
 const Container = styled.div`
   flex: 5 0 0;
@@ -94,46 +94,6 @@ const MessageItemBlock = styled.div`
   }
 `;
 
-const NoChannelInfoDiv = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-const NoChannelInfoDescription = styled.span`
-  margin-top: 20px;
-  font-size: 25px;
-`;
-
-const NoInfoImg = styled.img`
-  width: 150px;
-  height: 150px;
-  margin-bottom: 25px;
-`;
-
-const ResetButton = styled.button`
-  width: 250px;
-  height: 50px;
-  background: none;
-  padding: 15px 10px;
-  margin: 15px 0px 0px 0px;
-  border: 0;
-  outline: 0;
-  text-align: center;
-  vertical-align: middle;
-  border-radius: 10px;
-  background-color: #26a9ca;
-  cursor: pointer;
-  transition: all 0.3s;
-  &:hover {
-    background-color: #2dc2e6;
-    transition: all 0.3s;
-  }
-`;
-
 type MessageSectionProps = {
   messageList: MessageListInfo;
   setIsThreadOpen: React.Dispatch<React.SetStateAction<boolean>>;
@@ -142,7 +102,7 @@ type MessageSectionProps = {
 };
 
 function MessageSection(props: MessageSectionProps): JSX.Element {
-  const { selectedChannel, setSelectedMessageData, getServerChannelList, setIsModalOpen, setModalContents, socket } =
+  const { selectedChannel, setSelectedMessageData, setIsModalOpen, setModalContents, socket } =
     useContext(MainStoreContext);
   const { messageList, setIsThreadOpen, userList, channelInfo } = props;
   const { messageData } = messageList;
@@ -167,10 +127,6 @@ function MessageSection(props: MessageSectionProps): JSX.Element {
     setIsModalOpen(true);
   };
 
-  const onClickChannelListResetButton = () => {
-    getServerChannelList();
-  };
-
   const buildMessageItemList = () => {
     return messageData.map((val: MessageData): JSX.Element => {
       const { id } = val;
@@ -193,11 +149,7 @@ function MessageSection(props: MessageSectionProps): JSX.Element {
   return (
     <Container>
       {!channelInfo ? (
-        <NoChannelInfoDiv>
-          <NoInfoImg src={noInfoImg} />
-          <NoChannelInfoDescription>채널이 존재하지 않습니다...</NoChannelInfoDescription>
-          <ResetButton onClick={onClickChannelListResetButton}>새고로침</ResetButton>
-        </NoChannelInfoDiv>
+        <NotFoundChannel />
       ) : (
         <>
           <MessageSectionHeader>
